% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf_calc.R
\name{ddf_calc}
\alias{ddf_calc}
\title{Historical and difference DDF calculation}
\usage{
ddf_calc(ddf_dt)
}
\arguments{
\item{ddf_dt}{A processed dataset with historical RCP yet included}
}
\value{
A final dataset with all the data. Ready for plotting
}
\description{
Historical and difference DDF calculation
}
