% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf_precip.R
\name{ddf_precip}
\alias{ddf_precip}
\title{DDF for precipitation data}
\usage{
ddf_precip(dt)
}
\arguments{
\item{dt}{A dataset with p(precipitation kg/s), date and SID(filenames)}
}
\value{
A preprocessed dataset with new columns
}
\description{
DDF for precipitation data
}
