% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf_temp.R
\name{ddf_temp}
\alias{ddf_temp}
\title{DDF for temperature data}
\usage{
ddf_temp(dt)
}
\arguments{
\item{dt}{A dataset with TAS(temperature), DTM(date) and SID(filenames)}
}
\value{
A dataset with new columns
}
\description{
DDF for temperature data
}
