% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dates_forc.R
\name{read_Dates_forcing}
\alias{read_Dates_forcing}
\title{Reading dates and focing from filename}
\usage{
read_Dates_forcing(dir)
}
\arguments{
\item{dir}{A directory of the raw files}
}
\value{
A dataset with date and forcing columns
}
\description{
Reading dates and focing from filename
}
