% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{readData}
\alias{readData}
\title{Data reading function}
\usage{
readData(dir, na.rm = FALSE)
}
\arguments{
\item{dir}{dir() with all the raw data files is passed here}

\item{na.rm}{FALSE by default}
}
\value{
A one whole dataset with filename and data
}
\description{
Data reading function
}
\examples{

}
