% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_sp.R
\name{readData_sp}
\alias{readData_sp}
\title{Preprocessing spatial dataset}
\usage{
readData_sp(dir, ext)
}
\arguments{
\item{dir}{A directory with .nc files}

\item{ext}{Extention for the area}
}
\value{
A dataset with cells id as columnnames and precipitation data
}
\description{
Preprocessing spatial dataset
}
