% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_summary.R
\name{spatial_summary}
\alias{spatial_summary}
\title{Spatial data summary}
\usage{
spatial_summary(dir, aux, dates_forcing)
}
\arguments{
\item{dir}{A dataset directory}

\item{aux}{Dataset with cell's id and precipitation values}

\item{dates_forcing}{Dataset with dates and forcing}
}
\value{
A completed dataset
}
\description{
Spatial data summary
}
